+++ ConfigoSDK/ConfigoSDK/Constants/CFGConstants.h
@@ -41,5 +41,6 @@ typedef NS_ENUM(NSUInteger, CFGEnvironment) {
 + (NSURL *)eventsPushUrl;
 + (CFGEnvironment)currentEnvironment;
 
++ (NSError *)errorWithType:(CFGErrorCode)code userInfo:(NSDictionary *)info;
 
 @end