+++ ConfigoSDK/ConfigoSDK/Constants/CFGConstants.m
@@ -54,19 +54,16 @@ typedef NS_ENUM(NSUInteger, CFGApiVersion) {
 }
 
 + (NSString *)errorCodeToString:(CFGErrorCode)code {
-    NSString *retval = nil;
     switch(code) {
+        case CFGErrorNotConnected:
+            return @"notConnected";
         case CFGErrorBadResponse:
-            retval = @"badResponse";
-            break;
+            return @"badResponse";
         case CFGErrorRequestFailed:
-            retval = @"requestFailed";
-            break;
+            return @"requestFailed";
         default:
-            retval = @"unexpected";
-            break;
+            return @"unexpected";
     }
-    return retval;
 }
 
 #pragma mark - URL Builders